requirements.txt
charset-normalizer==3.3.2
idna==3.6
numpy==1.24.4
packaging==23.2
pandas==2.0.3
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
